backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.1.2
mysql-connector-python==8.3.0
pytest==7.4.3
hypothesis==6.92.1
